% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_obs_death_rates.R
\name{close_obs_death_rates}
\alias{close_obs_death_rates}
\title{Closing the observed death rations}
\usage{
close_obs_death_rates(deaths_obs, exp_obs, kannisto_nages, kannisto_nobs)
}
\arguments{
\item{deaths_obs}{The death counts.}

\item{exp_obs}{The exposures numbers.}

\item{kannisto_nages}{The extrapolation ages used for closing the observed mortality
rates.}

\item{kannisto_nobs}{The number of observations used to extrapolate.}
}
\description{
This functions closes the observed death rates of the country of interest.
}
\examples{
lst   <- MultiMoMo::european_mortality_data
dat_M <- lst$M
dat_F <- lst$F

deaths_obs <- dat_M$UNI$BE$dtx
exp_obs    <- dat_M$UNI$BE$etx
kannisto_nages <- 30
kannisto_nobs  <- 11
close_obs_m    <- close_obs_death_rates(deaths_obs, exp_obs, kannisto_nages, kannisto_nobs)
}
\keyword{internal}
