% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_li_lee.R
\name{fit701}
\alias{fit701}
\title{Newton-Raphson implementation to fit a single Lee-Carter mortality model (M1)}
\usage{
fit701(xv, yv, etx, dtx, wa, int, constraints, exclude_coh)
}
\arguments{
\item{xv}{The vector of ages.}

\item{yv}{The vector of years.}

\item{etx}{m x n matrix of exposures.}

\item{dtx}{m x n matrix of death counts.}

\item{wa}{m x n matrix of weights (0 or 1).}

\item{int}{The prespecified vector for \eqn{beta1(x)}. When the zero-vector is supplied,
the Newton_Raphson algorithm calibrates \eqn{beta1(x)}}

\item{constraints}{Character string. If \code{constraints = "ALL"}, two
identifiability constraints are imposed (see details). If
\code{constraints == "ADJUST"}, only the constraint on beta2(x) is used.}

\item{exclude_coh}{logical. If TRUE, the weights are set to zero for cohorts with
fewer than 5 observations.}
}
\value{
A list containing the fitted parameter estimates of the type M1 Lee-Carter model, as
well as some model specifications.
}
\description{
This internal function fits a Lee-Carter mortality model (type M1) with the use
of Newton-Raphson. This mortality model M1 has the following form:
\deqn{log m(t,x) = beta1(x) + beta2(x)*kappa2(t) + Poisson error}
}
\details{
The two identifiability constraints are:
\deqn{kappa2(t1) = 0,    sum(beta2(x)^2) = 1.}
The BIC is defined as:
\deqn{-2*l1 + log(sum(wa)),}
and where the log-likelihood equals:
\deqn{sum((dtx*log(etx*mhat.tx) - etx*mhat.tx - lgamma(dtx + 1))*wa).}
Here, \eqn{mhat.tx} denotes the fitted forces of mortality, \eqn{etx} the
exposure numbers and \eqn{dtx} the death counts.
}
\keyword{internal}
