% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_li_lee.R
\name{fit_li_lee}
\alias{fit_li_lee}
\title{Fit a multipopulation Li-Lee mortality model}
\usage{
fit_li_lee(xv, yv, yv_spec, country_spec, data, method, Ax, exclude_coh)
}
\arguments{
\item{xv}{The vector of ages.}

\item{yv}{The vector of years.}

\item{yv_spec}{The vector of years for the country of interest.}

\item{country_spec}{The country of interest.}

\item{data}{The mortality data (dtx, wtx, wa) for each country.}

\item{method}{The method to fit the mortality model (currently only \code{"NR"} working).}

\item{Ax}{Logical. Should a common age specific factor be included in the model?}

\item{exclude_coh}{Logical. If TRUE, the weights are set to zero for cohorts with
fewer than 5 observations.}
}
\value{
A list containing the following objects
\itemize{
\item Common parameters: $Ax, $Bx and $Kt
\item Country specific parameters: $ax, $bx and $kt
\item The force of mortality estimates: $mtx
\item The log-likelihood of the model: $ll
\item The number of parameters in the model: $npar
\item The BIC of the model: $BIC
\item The Poisson Pearson residuals of the mortality model: $residuals
}
}
\description{
This function fits a multipopulation Li-Lee model. This model consists of two times a
Lee & Carter model: one for the common trend and one for the country-specific
deviation from this trend. The model has the following form
\describe{
\item{}{\eqn{log(mtx) = log(mtx^{(T)}) + log(mtx^{(c)})}}
\item{}{\eqn{log(mtx^{(T)}) = Ax + Bx * Kt}}
\item{}{\eqn{log(mtx^{(c)}) = ax + bx * kt},}
}
where \eqn{log(mxt)} refers to the force of mortality for an x-year old at time t.
}
\details{
The \code{data} must be (in the same format as) the output of the function
\code{\link[MultiMoMo]{get_mortality_data}}.

The parameters are fitted using a two step Newton-Raphson approach.
In the first step, we fit the parameters for the commond trend
(\eqn{Ax}, \eqn{Bx} and \eqn{Kt}) (standard Lee-Carter fit).
In a second step, we keep \eqn{Ax}, \eqn{Bx} and \eqn{Kt} fixed and assume
that \deqn{dtx ~ POI(etx*mtx^{(T)}*mtx^{(c)}),} where \eqn{mtx^{(T)}}
denote the fitted forces of mortality for the common trend, namely
\deqn{mtx^{(T)} = exp(Ax + Bx*Kt),}
and where the \eqn{mtx^{(c)}} denote the fitted forces of mortality
for the country-specific deviation from the common trend, namely
\deqn{mtx^{(c)} = exp(ax + bx* kt).}
So in the second step, we again fit a standard Lee-Carter model for the deaths
of your country of interest \eqn{dtx} and with the adapted exposures \eqn{etx*mtx^{(T)}}.

The BIC is defined as:
\deqn{BIC = -2*ll + log(sum(wa)).}

The log-likelihood \eqn{ll} equals:
\deqn{ll = sum((dtx*log(etx*mtx) - etx*mtx - lgamma(dtx + 1))*wa).}
Here, \eqn{mtx} denotes the fitted forces of mortality for your country of interest,
\eqn{etx} its exposure numbers and \eqn{dtx} the death counts.

The Poisson Pearson residuals equal:
\deqn{(dtx - etx*mtx)/sqrt(etx*mtx).}
}
\examples{
lst   <- MultiMoMo::european_mortality_data
dat_M <- lst$M
dat_F <- lst$F
xv    <- 0:90
yv = yv_spec <- 1970:2018
Countries   <- names(dat_M$UNI)
country_spec <- "BE"
fit_M <- fit_li_lee(xv, yv, yv_spec, country_spec, dat_M, "NR", TRUE, FALSE)
fit_F <- fit_li_lee(xv, yv, yv_spec, country_spec, dat_F, "NR", TRUE, FALSE)

}
