% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_codes.R
\name{get_country_codes}
\alias{get_country_codes}
\title{Get the available countries within the package}
\usage{
get_country_codes(data_magec = NULL, data_mager = NULL, data_pjan = NULL)
}
\arguments{
\item{data_magec}{The dataset of the period death counts on Eurostat. If \code{NULL},
this dataset is downloaded from \href{https://ec.europa.eu/eurostat/}{Eurostat}.}

\item{data_mager}{The dataset of the cohort death counts on Eurostat. If \code{NULL},
this dataset is downloaded from \href{https://ec.europa.eu/eurostat/}{Eurostat}.}

\item{data_pjan}{The dataset of the population numbers at 1 January of each year on Eurostat. If \code{NULL},
this dataset is downloaded from \href{https://ec.europa.eu/eurostat/}{Eurostat}.}
}
\value{
A dataframe containing the available country names, the corresponding HMD, Eurostat
and user country labels as well as the avaible year range for each country.
}
\description{
This function shows the available countries that can be used in the
multipopulation mortality model. Further, their user codes are given as well as the
available year range for each country. Information is substracted from the data sources
\href{https://www.mortality.org/}{HMD} and \href{https://ec.europa.eu/eurostat/}{Eurostat}.
}
\details{
This function can be used to find your desired country set and year range to construct
the multipopulation mortality model.

A special case: Germany. Here before 1990 (German unification), only data from West-Germany
is considered.
}
\examples{
df_country <- get_country_codes()

}
