% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_specific_range.R
\name{get_data_specific_range}
\alias{get_data_specific_range}
\title{Adapt the year and/or age range of the multi-population mortality dataset}
\usage{
get_data_specific_range(xv, yv, yv_spec, data, country, country_spec)
}
\arguments{
\item{xv}{The new vector of ages.}

\item{yv}{The new vector of years.}

\item{yv_spec}{The new vector of years for the country of interest.}

\item{data}{The multi-population mortality dataset.}

\item{country}{The vector of countries.}

\item{country_spec}{The country of interest.}
}
\value{
A list containing the following updated objects:
\itemize{
\item $M (male data), $F (female data)
\item $UNI (individual mortality data: $BE (Belgium), $NL (Netherlands), $AT (Austria), ...), $ALL (combined data)
\item $dtx (deaths), $etx (exposures), $wa (weights).
}
}
\description{
This function adapts the year and/or age range of the mortality dataset,
as obtained from the function \code{\link[MultiMoMo]{get_mortality_data}},
for each country in the multi-population model.
}
\details{
The argument \code{data} must be (in the same form as) the output
from the function \code{\link[MultiMoMo]{get_mortality_data}}. See
\code{\link[MultiMoMo]{european_mortality_data}} for such an example mortality
dataset.
}
\examples{
xv       <- 0:90
yv       <- 1988:2018
yv_spec  <- 1988:2018
data     <- MultiMoMo::european_mortality_data
country  <- names(data$M$UNI)
country_spec <- "BE"

data_new <- get_data_specific_range(xv, yv, yv_spec, data, country, country_spec)

}
