% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mortality_data.R
\name{get_mortality_data}
\alias{get_mortality_data}
\title{Get the multipopulation mortality dataset}
\usage{
get_mortality_data(
  xv,
  yv,
  yv_spec,
  countries,
  country_spec,
  username,
  password
)
}
\arguments{
\item{xv}{Numeric. The vector of ages.}

\item{yv}{The vector of years.}

\item{yv_spec}{The vector of years for the country of interest.}

\item{countries}{The vector of countries.}

\item{country_spec}{The country of interest.}

\item{username}{The username of your HMD account.}

\item{password}{The password of your HMD acccount.}
}
\value{
A list containing:
\itemize{
\item $M (male data), $F (female data)
\item $UNI (individual mortality data: $BE (Belgium), $NL (Netherlands), $AT (Austria), ...), $ALL (combined data)
\item $dtx (deaths), $etx (exposures), $wa (weights).
}
}
\description{
This function returns the country-specific and combined \verb{death counts} and \code{exposures} for a selection
of countries from the Human Mortality Database (\href{https://www.mortality.org/}{HMD}) and/or
\href{https://ec.europa.eu/eurostat/}{Eurostat} and for both males and females.
}
\details{
This function downloads and/or calculates the death counts and exposures for an
age range \code{xv}, a time range \code{yv} and for a specified set of countries \code{countries}.
The main data source is HMD. The second one is Eurostat for the years for which there is no
data available at the HMD. From the HMD database this functions downloads the tables
\code{Deaths} and \verb{Exposure to risk} in 1x1 format. In the case of Germany (\code{DE}), we consider
West-Germany instead of whole Germany for the period before the German unification
(...-1989). At the Eurostat database we download the databases \code{demo_mager} (period deaths),
\code{demo_magec} (cohort deaths) and \code{demo_pjan} (population size). Period exposures are
calculated from these datasets according to the HMD protocol.

The argument \code{countries} should contain the user code labels of your countries of
interest. The available countries and their corresponding \verb{user codes} are given in the dataset
\code{\link{country_codes}}, provided in this package.

Further, you have to register on \href{https://www.mortality.org/}{HMD} and provide
the function with your \code{username} and \code{password} such that the function can
automatically download from HMD.

Remark that the year ranges \code{yv} and \code{yv_spec} should not be the same. Typically
a longer year range for yv is chosen and yv_spec consists of some more recent years.
When we download data for your country of interest (\code{country_spec}), we consider
the year range \code{min(yv,yv_spec):tail(yv_spec,1)} since the older data is also used for estimating
the common mortality trend.
}
\examples{
\dontrun{
xv <- 0:90
yv <- 1970:2015
yv_spec <- 1985:2018
countries <- c("BE", "IT", "UK", "DE")
country_spec <- "BE"
username <- ""
password <- ""
data <- get_mortality_data(xv, yv, yv_spec, countries, country_spec, username, password)
}

}
