% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_eurostat.R
\name{mapvalues}
\alias{mapvalues}
\title{Function mapvalues from tidyr pakcage}
\usage{
mapvalues(x, from, to, warn_missing = TRUE)
}
\description{
Function mapvalues from tidyr pakcage
}
\keyword{internal}
