% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e0_hmd.R
\name{plot_e0_hmd}
\alias{plot_e0_hmd}
\title{The evolution of the life expectancy of a newborn}
\source{
\\href{https://www.mortality.org/}
}
\usage{
plot_e0_hmd(Country, username, password)
}
\arguments{
\item{Country}{The set of countries.}

\item{username}{The username of your HMD account.}

\item{password}{The password of your HMD account.}
}
\description{
Graph of the evolution of the life expectancy of a male and female newborn.
Data is downloaded from the Human Mortality Database (\href{https://www.mortality.org/}{HMD}).
Therefore, only countries available in the HMD are visualized in this graph.
}
\examples{
\dontrun{
  Country   <- c("BE", "IT", "DE", "FR", "NL")
  username  <- ""
  password  <- ""
  plot_e0_hmd(Country, username, password)}

}
\keyword{ggplots}
