% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameters_li_lee.R
\name{plot_parameters_li_lee}
\alias{plot_parameters_li_lee}
\title{Plot the parameters of the multipopulation (Li-Lee) model}
\usage{
plot_parameters_li_lee(xv, yv, fit, sex, CountrySPEC, method, type)
}
\arguments{
\item{xv}{The vector of ages.}

\item{yv}{The vector of years (for your country of interest).}

\item{fit}{The estimated Li-Lee model.}

\item{sex}{The gender of interest (\code{"Male"} or \code{"Female"})}

\item{CountrySPEC}{The user code of your country of interest.}

\item{method}{The method used to fit the Li-Lee mortality model (currently only \code{"NR"} working).}

\item{type}{Which type of parameter estimates should be visualized? Either one of the common
parameter estimates (\code{"A.x"}, \code{"B.x"} or \code{"K.t"}), one of the country-specific
estimates (\code{"a.x"}, \code{"b.x"} or \code{"k.t"}) or all of them combined in one plot (\code{"ALL"})}
}
\value{
A plot showing the parameter estimates of the fitted Li-Lee model.
}
\description{
This function plots the estimated parameters in the multipopulation Li-Lee model.
}
\details{
The object \code{fit} must be the output of the function \code{\link{fit_li_lee}}.
}
\examples{
lst   <- MultiMoMo::european_mortality_data
dat_M <- lst$M
xv    <- 0:90
yv = yvSPEC <- 1970:2018
Countries   <- names(dat_M$UNI)
CountrySPEC <- "BE"
fit_M <- fit_li_lee(xv, yv, yvSPEC, CountrySPEC, dat_M, "NR", TRUE, FALSE)
plot_parameters_li_lee(xv, yvSPEC, fit_M, "Male", CountrySPEC, "NR", "ALL")

}
