% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_li_lee.R
\name{plot_residuals_li_lee}
\alias{plot_residuals_li_lee}
\title{Plot the Poisson Pearson residuals of the multipopulation Li-Lee model}
\usage{
plot_residuals_li_lee(xv, yv, fit, CountrySPEC, sex)
}
\arguments{
\item{xv}{The vector of ages.}

\item{yv}{The vector of years (for your country of interest).}

\item{fit}{The fitted Li-Lee model.}

\item{CountrySPEC}{The user code of your country of interest.}

\item{sex}{The gender of interest (\code{"Male"} or \code{"Female"})}
}
\value{
A level plot showing the Poisson Pearson residuals.
}
\description{
This function returns a level plot of the Poisson Pearson residuals
of the fitted multipopulation Li-Lee model. These Pearson residuals are defined as
\deqn{(d_{x,t} - \hat{mu}_{x,t}E_{x,t})/\sqrt{\hat{mu}_{x,t}E_{x,t}},}
and are visualized as a function of age (\code{xv}) and time (\code{yv}).
The greener the Pearson residuals are coloured, the closer they are to zero
and hence the better the estimated deaths from the Li-Lee model correspond
to the observed deaths.
}
\details{
The object \code{fit} must be the output of the function \code{\link{fit_li_lee}}.
}
\examples{
lst   <- MultiMoMo::european_mortality_data
dat_M <- lst$M
xv    <- 0:90
yv = yvSPEC <- 1970:2018
Countries   <- names(dat_M$UNI)
CountrySPEC <- "BE"
fit_M <- fit_li_lee(xv, yv, yvSPEC, CountrySPEC, dat_M, "NR", TRUE, FALSE)
plot_residuals_li_lee(xv, yvSPEC, fit_M, CountrySPEC, "Male")

}
