% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mortality_rates.R
\name{project_mortality_rates}
\alias{project_mortality_rates}
\title{Project mortality rates}
\usage{
project_mortality_rates(fit_M, fit_F, proj_par)
}
\arguments{
\item{fit_M}{The calibrated Li-Lee model for males.}

\item{fit_F}{The calibrated Li-Lee model for females.}

\item{proj_par}{The (simulations of) the projected time dependent parameters
\mjeqn{(K_t^M, \kappa_t^M, K_t^F, \kappa_t^F)}{ASCII representation}.}
}
\value{
A list containing the following objects
\itemize{
\item The projected/simulated male mortality rates ($Male).
\item The projected/simulated female mortality rates ($Female).
}
}
\description{
This function project mortality rates using the simulations that
were set up for the joint time series \mjeqn{(K_t^M, \kappa_t^M, K_t^F,
\kappa_t^F)}{ASCII representation}.
}
\details{
\loadmathjax

The arguments \code{fit_M} and \code{fit_F} should be the outputs of the function
\code{\link{fit_li_lee}}.

The argument \code{proj_par} should be the output of the function
\code{\link{project_parameters}}.
}
\examples{
lst   <- MultiMoMo::european_mortality_data
dat_M <- lst$M
dat_F <- lst$F
xv    <- 0:90
yv = yvSPEC <- 1970:2018
Countries   <- names(dat_M$UNI)
CountrySPEC <- "BE"
fit_M <- fit_li_lee(xv, yv, yvSPEC, CountrySPEC, dat_M, "NR", TRUE, FALSE)
fit_F <- fit_li_lee(xv, yv, yvSPEC, CountrySPEC, dat_F, "NR", TRUE, FALSE)

arima_spec <- list(K.t_M = "RWD", k.t_M = "AR3.1", K.t_F = "RWD", k.t_F = "AR5.0")
n_ahead    <- 50
n_sim      <- 10000
est_method <- "PORT"
proj_par   <- project_parameters(fit_M, fit_F, n_ahead, n_sim, arima_spec, est_method)
proj_rates <- project_mortality_rates(fit_M, fit_F, proj_par)


}
