% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_parameters.R
\name{project_parameters}
\alias{project_parameters}
\title{Estimate the time series of the time dependent parameters in the calibrated Li-Lee model and
make projections}
\usage{
project_parameters(fit_M, fit_F, n_ahead, n_sim, arima_spec, est_method)
}
\arguments{
\item{fit_M}{The calibrated Li-Lee model for males.}

\item{fit_F}{The calibrated Li-Lee model for females.}

\item{n_ahead}{The number of years to project in the future.}

\item{n_sim}{The number of simulations/projections.}

\item{arima_spec}{The ARIMA time series specifications for \mjeqn{(K_t^M, \kappa_t^M, K_t^F, \kappa_t^F)}{ASCII representation}.}

\item{est_method}{The estimation method of the parameters in the time series specifications.}
}
\value{
A list containing the following objects
\itemize{
\item The estimated parameters in the time series for \mjeqn{K_t^M}{ASCII representation}:
$coef_KtM
\item The estimated parameters in the time series for \mjeqn{\kappa_t^M}{ASCII representation}:
$coef_ktM
\item The estimated parameters in the time series for \mjeqn{K_t^F}{ASCII representation}:
$coef_KtF
\item The estimatedparameters in the time series for \mjeqn{\kappa_t^F}{ASCII representation}:
$coef_ktF
\item The estimated covariance matrix \mjeqn{C}{ASCII representation}:
$cov_mat
\item The \mjeqn{K_t^M}{ASCII representation} simulations: $K.t_M
\item The \mjeqn{\kappa_t^M}{ASCII representation} simulations: $k.t_M
\item The \mjeqn{K_t^F}{ASCII representation} simulations: $K.t_F
\item The \mjeqn{\kappa_t^F}{ASCII representation} simulations: $k.t_F
}
}
\description{
This function estimates the parameters in the  time series specifications
for the time dependent parameters in the calibrated Li-Lee model, namely
\mjeqn{K_t}{ASCII representation} and \mjeqn{\kappa_t}{ASCII representation},
jointly for males and females.
}
\details{
\loadmathjax

The arguments \code{fit_M} and \code{fit_F} are the outputs of the function
\code{\link{fit_li_lee}}.

The argument \code{arima_spec} is a list specifying the type of time series used
for each time dependent parameter in the Li-Lee model and is of the form:
\deqn{list("K.t_M" = , "k.t_M" =  , "K.t_F" = , "k.t_F" = ).}
The capital \code{K.t} refers to the time dependent parameter in the Lee-Carter model
for the common trend, i.e. \mjeqn{K_t}{ASCII representation}. The small \code{k.t} refers
to the time dependent parameter \mjeqn{\kappa_t}{ASCII representation}
in the Lee-Carter model for the country-specific deviation from the common trend.
We jointly model the time series dynamics for men and women by assuming a
multivariate Gaussian distribution with mean \mjeqn{(0,0,0,0)}{ASCII represenation} and covariance matrix
\mjeqn{C}{ASCII represenation} for the error terms
\mjeqn{(\epsilon_t^M, \delta_t^M, \epsilon_t^F, \delta_t^F)}{ASCII representation}
of the multivariate time series \mjeqn{(K_t^M, \kappa_t^M, K_t^F, \kappa_t^F)}{ASCII representation}.
Possible choices for the time series dynamics are \code{RWD} (e.g. for \mjeqn{K_t}{ASCII representation})
or \code{ARk.j}, where \code{k} refers to an AR(k) process, an auto-regressive process
with lag \code{k}. Further \mjeqn{j \in \{0,1\}}{ASCII representation} and refers
to an AR process without or with an intercept respectively.

The argument \code{est_method} specifies the estimation method of the parameters in
the time series specifications, given in \code{arima_spec}. A first option is \code{SUR},
referring to seemingly unrelated regression, that uses the function \code{\link[systemfit]{systemfit}}.
This method only works for a limited amount of cases. In particular, all time series
must jointly start at the same time t. For example, this means that in the example below,
we start the joint time series estimation at year 1975 (since an \code{AR5.0} is used).
A second option is \code{PORT}. This is a self-written objective function that
maximizes the log-likelihood of the multivariate Gaussian distribution for
\mjeqn{(\epsilon_t^M, \delta_t^M, \epsilon_t^F, \delta_t^F)}{ASCII represenation} using the function
\code{\link[stats]{nlminb}} in the \code{stats}-package. In the example below, the joint estimation
consists of three parts. For t = 1971-1972 only the \code{RWD} processes are considered leading to
a bi-variate normal log-likelihood. Then for t = 1973-1974 the two \code{RWD} processes together with
\code{k.t_M} are used, leading to a 3-variate normal log-likelihood. From 1975 on, we deal with the
four processes (4-variate normal log-likelihood). The sum of these three likelihoods is maximized
at once using PORT routines.
}
\examples{
lst   <- MultiMoMo::european_mortality_data
dat_M <- lst$M
dat_F <- lst$F
xv    <- 0:90
yv = yvSPEC <- 1970:2018
Countries   <- names(dat_M$UNI)
CountrySPEC <- "BE"
fit_M <- fit_li_lee(xv, yv, yvSPEC, CountrySPEC, dat_M, "NR", TRUE, FALSE)
fit_F <- fit_li_lee(xv, yv, yvSPEC, CountrySPEC, dat_F, "NR", TRUE, FALSE)

arima_spec <- list(K.t_M = "RWD", k.t_M = "AR3.1", K.t_F = "RWD", k.t_F = "AR5.0")
n_ahead    <- 50
n_sim      <- 10000
est_method <- "PORT"
proj       <- project_parameters(fit_M, fit_F, n_ahead, n_sim, arima_spec, est_method)

}
